requirements.txt
# gym==0.21.0
gymnasium==0.29.1
stable_baselines3==1.1.0
tqdm==4.51.0
opencv_python==4.5.1.48
matplotlib==3.4.2
numpy==1.19.5
Pillow==8.4.0
gym-simplifiedtetris
torch~=1.10.0
opencv-python~=4.5.1.48